requirements.txt
boto==2.49.0
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.42.0
django-ses==1.0.1
docutils==0.15.2
future==0.18.2
idna==2.10
jmespath==0.10.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-twitter==3.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
